moved the command to prefix.js and renamed it to parsers.js
