+++ src/commands.js
@@ -409,114 +409,4 @@ Commands.leaderboard = new Command("See the users with the highest ranks", (mess
     message.channel.send(LeaderBoard)
 }, "Economy")
 
-aliases = new class extends Discord.Collection {
-    constructor() {
-        super();
-        this.reload();
-    }
-    reload() {
-        const object = JSON.parse(fs.readFileSync("./src/Data/aliases.json", "utf-8"));
-
-        Object.entries(object).forEach(([key, value]) => {
-            this.set(key, value);
-        });
-    }
-    save() {
-        const json = JSON.stringify(Object.fromEntries(this), null, 4);
-        fs.writeFileSync("./src/Data/aliases.json", json, "utf8");
-    }
-}
-const aliasHelp = `
-\`&alias set (name) (substitute)\` set an alias for \`substitute\` with the name \`name\`
-\`&alias get (name)\` get an alias with the name \`name\`
-\`&alias delete (name)\` delete an alias with the name \`name\`
-\`&alias clear\` clear all alias
-\`&alias list\` list all alias
-`.trim();
-
-Commands.alias = new Command("Manage command aliases", (message, [command, ...args]) => {
-    switch (command) {
-        case "set": {
-            const [name, substitute] = args;
-
-            if (name in Commands)
-                return void message.channel.send("error: an existing command already has the specified name");
-
-            if (name == undefined)
-                return void message.channel.send("error: missing alias name");
-
-            if (substitute == undefined)
-                return void message.channel.send("error: missing alias value");
-
-            aliases.set(message.author.id, {
-                ...aliases.get(message.author.id),
-                [name]: String(substitute)
-            }).save();
-
-            message.channel.send("alias set successfully");
-            break;
-        }
-        case "get": {
-            const [name] = args;
-
-            if (name == undefined)
-                return void message.channel.send("error: missing alias name");
-
-            if (!(name in (aliases.get(message.author.id) ?? {})))
-                return void message.channel.send("error: invalid alias name");
-
-            message.channel.send(
-                '```properties\n' +
-                name + ' = ' + aliases.get(message.author.id)[name]
-                + '\n```'
-                ?? "error: missing invalid name");
-            break;
-        }
-        case "delete": {
-            const [name] = args;
-
-            if (name == undefined)
-                return void message.channel.send("error: missing alias name");
-
-            if (!(name in (aliases.get(message.author.id) ?? {})))
-                return void message.channel.send("error: invalid alias name");
-
-            delete aliases.get(message.author.id)[name]
-            aliases.save();
-            message.channel.send("alias deleted successfully");
-            break;
-        }
-        case "clear":
-            aliases.delete(message.author.id)
-            aliases.save();
-            message.channel.send("aliases cleared successfully");
-            break;
-        case "list":
-            message.channel.send(
-                new Discord.MessageEmbed()
-                    .setTitle(`${message.author.username}'s aliases`)
-                    .setDescription(
-                        Object.keys(aliases.get(message.author.id) ?? {}).length > 0 ?
-                            '```properties\n' +
-                            Array.from(Object.entries(aliases.get(message.author.id) ?? {}))
-                                .map(([name, substitute]) =>
-                                    `${name} = ${/[^A-Za-z0-9_$]/.test(substitute)
-                                        ? '"' + substitute + '"'
-                                        : substitute
-                                    }`
-                                ).join('\n')
-                            + '\n```' : '```\n<empty>\n```'
-                    )
-                    .setFooter(`${Object.keys(aliases.get(message.author.id) ?? {}).length} aliases`)
-            )
-            break;
-        default:
-            message.channel.send(aliasHelp.replace(/&/g, Prefix.get(message.guild.id)))
-            break
-    }
-}, "Utility", [
-    new RequiredArg(0, undefined, "command", true),
-    new RequiredArg(1, undefined, "argument", true),
-])
-
 require("./xilefunds")
